requirements.txt
python-telegram-bot
python-dotenv
pyautogui
psutil
Pillow
opencv-python
pytesseract
watchdog
